Move services to cswm.App project
