+++ src/cswm.sln
@@ -13,6 +13,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "cswm.Tests", "..\test\cswm.Tests\cswm.Tests.csproj", "{45C5A52A-CF9E-4F53-BE98-D501EADA3CBA}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "cswm.App", "cswm.App\cswm.App.csproj", "{1AAB59C0-5A2B-4304-ABFF-491C44D06D6C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{45C5A52A-CF9E-4F53-BE98-D501EADA3CBA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{45C5A52A-CF9E-4F53-BE98-D501EADA3CBA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{45C5A52A-CF9E-4F53-BE98-D501EADA3CBA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1AAB59C0-5A2B-4304-ABFF-491C44D06D6C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1AAB59C0-5A2B-4304-ABFF-491C44D06D6C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1AAB59C0-5A2B-4304-ABFF-491C44D06D6C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1AAB59C0-5A2B-4304-ABFF-491C44D06D6C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
