--- src/cswm/Properties/launchSettings.json
@@ -1,17 +1,0 @@
-{
-  "profiles": {
-    "cswm": {
-      "commandName": "Project",
-      "environmentVariables": {
-        "DOTNET_ENVIRONMENT": "Development"
-      }
-    },
-    "cswm - Do Not Manage": {
-      "commandName": "Project",
-      "commandLineArgs": "--WindowManagementOptions:DoNotManage=true",
-      "environmentVariables": {
-        "DOTNET_ENVIRONMENT": "Development"
-      }
-    }
-  }
-}
