--- src/cswm/Services/IService.cs
@@ -1,7 +1,0 @@
-namespace cswm.Services;
-
-public interface IService
-{
-    public void Start();
-    public void Stop();
-}
