--- src/cswm/appsettings.json
@@ -1,10 +1,0 @@
-{
-  "WindowManagementOptions": {
-    "MonitorPadding": 16,
-    "WindowMargin": 16,
-    "WindowPadding": 0,
-    "IgnoredWindowClasses": [
-      "ApplicationFrameWindow"
-    ]
-  }
-}
