+++ src/cswm/cswm.csproj
@@ -2,25 +2,11 @@
 
   <PropertyGroup>
     <AssemblyName>cswm</AssemblyName>
-    <Version>0.4.0</Version>
-    <ApplicationIcon>icon.ico</ApplicationIcon>
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>net6.0-windows</TargetFramework>
-    <UseWindowsForms>true</UseWindowsForms>
+    <TargetFramework>net6.0</TargetFramework>
     <Nullable>enable</Nullable>
-    
-    <!-- Single-file deployment -->
-    <!-- <https://learn.microsoft.com/en-us/dotnet/core/deploying/single-file/overview> -->
-    <PublishSingleFile>true</PublishSingleFile>
-    <SelfContained>true</SelfContained>
-    <RuntimeIdentifier>win-x64</RuntimeIdentifier>
-    <DebugType>embedded</DebugType>
-    <PublishReadyToRun>true</PublishReadyToRun>
-    <IncludeNativeLibrariesForSelfExtract>true</IncludeNativeLibrariesForSelfExtract>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.1" />
     <PackageReference Include="Microsoft.Extensions.Logging.Configuration" Version="6.0.0" />
     <PackageReference Include="Microsoft.Windows.CsWin32" Version="0.2.46-beta">
       <PrivateAssets>all</PrivateAssets>
@@ -29,25 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <EmbeddedResource Include="icon.ico">
-      <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
-    </EmbeddedResource>
-  </ItemGroup>
-
-  <ItemGroup>
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-      <_Parameter1>$(AssemblyName).Tests</_Parameter1>
-    </AssemblyAttribute>
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="appsettings.Development.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      <CopyToPublishDirectory>Never</CopyToPublishDirectory>
-    </None>
-    <None Update="appsettings.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
+    <InternalsVisibleTo Include="cswm.Tests" />
   </ItemGroup>
 
 </Project>
