+++ test/cswm.Tests/cswm.Tests.csproj
@@ -23,7 +23,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\cswm\cswm.csproj" />
+    <ProjectReference Include="..\..\src\cswm\cswm.csproj">
+      <ReferenceSourceTarget></ReferenceSourceTarget>
+    </ProjectReference>
   </ItemGroup>
 
 </Project>
